fix: fixed import notations for Header component
