+++ src/app/layout.tsx
@@ -1,8 +1,7 @@
-import Header from '@/components/Layout/Header'
 import './globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
-import { Footer } from '@/components'
+import { Footer, Header } from '@/components'
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -17,7 +16,7 @@ export default function RootLayout({
   children: React.ReactNode
 }) {
   return (
-    <html lang="en">
+    <html className='dark' lang="en">
       <body className={inter.className}>
         <Header />
           <div className="site">
