+++ src/components/Header/styles.css
@@ -39,14 +39,14 @@
 }
 
 .headerColors {
-    background-color: #604f70a1;
+    background-color: #5c0baba1;
     -webkit-transition: box-shadow 300ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
     transition: box-shadow 300ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-    box-shadow: 0px 2px 4px -1px rgba(0, 0, 0, 0.2), 0px 4px 5px 0px rgba(0, 0, 0, 0.14), 0px 1px 10px 0px rgba(0, 0, 0, 0.12);
+    box-shadow: 0px 0.1px 4px -2px rgba(0, 0, 0, 0.05), 0px 4px 5px 0px rgba(0, 0, 0, 0.05), 0px 1px 10px 0px rgba(0, 0, 0, 0.05);
 }
 
 .logobody {
-    background-color: #604f70a1;
+    background-color: #5c0baba1;
     border-radius: 10px;
     display: flex;
     padding: 0.5rem 1rem;
