+++ src/components/index.tsx
@@ -5,3 +5,4 @@ export { EventCard } from "./EventCard";
 export { WhatWeDoCard } from "./WhatWeDoCard";
 export { WhatWeDoSection } from "./WhatWeDoSection";
 export { Footer } from "./Footer";
+export { Header } from "./Header";
